+++ lib/stream/manager/api.go
@@ -1,6 +1,7 @@
 package manager
 
 import (
+	"context"
 	"encoding/json"
 	"errors"
 	"fmt"
@@ -11,11 +12,15 @@ import (
 	"sync"
 	"time"
 
+	"github.com/Jeffail/benthos/v3/internal/bundle"
 	"github.com/Jeffail/benthos/v3/internal/docs"
 	"github.com/Jeffail/benthos/v3/lib/buffer"
+	"github.com/Jeffail/benthos/v3/lib/cache"
 	"github.com/Jeffail/benthos/v3/lib/input"
 	"github.com/Jeffail/benthos/v3/lib/output"
 	"github.com/Jeffail/benthos/v3/lib/pipeline"
+	"github.com/Jeffail/benthos/v3/lib/processor"
+	"github.com/Jeffail/benthos/v3/lib/ratelimit"
 	"github.com/Jeffail/benthos/v3/lib/stream"
 	"github.com/Jeffail/benthos/v3/lib/util/text"
 	"github.com/Jeffail/gabs/v2"
@@ -44,6 +49,11 @@ func (m *Type) registerEndpoints() {
 		"/streams/{id}/stats",
 		"GET a structured JSON object containing metrics for the stream.",
 		m.HandleStreamStats,
+	)
+	m.manager.RegisterEndpoint(
+		"/resources/{type}/{id}",
+		"POST: Create or replace a given resource configuration of a specified type. Types supported are `cache`, `input`, `output`, `processor` and `rate_limit`.",
+		m.HandleResourceCRUD,
 	)
 	m.manager.RegisterEndpoint(
 		"/ready",
@@ -275,10 +285,11 @@ func (m *Type) HandleStreamCRUD(w http.ResponseWriter, r *http.Request) {
 		if confBytes, err = ioutil.ReadAll(r.Body); err != nil {
 			return
 		}
+		confBytes = text.ReplaceEnvVariables(confBytes)
 
 		if r.URL.Query().Get("chilled") != "true" {
 			var node yaml.Node
-			if requestErr = yaml.Unmarshal(confBytes, &node); requestErr != nil {
+			if err = yaml.Unmarshal(confBytes, &node); err != nil {
 				return
 			}
 			lints = lintStreamConfigNode(&node)
@@ -288,7 +299,7 @@ func (m *Type) HandleStreamCRUD(w http.ResponseWriter, r *http.Request) {
 		}
 
 		confOut = stream.NewConfig()
-		err = yaml.Unmarshal(text.ReplaceEnvVariables(confBytes), &confOut)
+		err = yaml.Unmarshal(confBytes, &confOut)
 		return
 	}
 	patchConfig := func(confIn stream.Config) (confOut stream.Config, err error) {
@@ -408,6 +419,129 @@ func (m *Type) HandleStreamCRUD(w http.ResponseWriter, r *http.Request) {
 		serverErr = nil
 		http.Error(w, "Stream already exists", http.StatusBadRequest)
 	}
+}
+
+// HandleResourceCRUD is an http.HandleFunc for performing CRUD operations on
+// resource components.
+func (m *Type) HandleResourceCRUD(w http.ResponseWriter, r *http.Request) {
+	var serverErr, requestErr error
+	defer func() {
+		if r.Body != nil {
+			r.Body.Close()
+		}
+		if serverErr != nil {
+			m.logger.Errorf("Resource CRUD Error: %v\n", serverErr)
+			http.Error(w, fmt.Sprintf("Error: %v", serverErr), http.StatusBadGateway)
+		}
+		if requestErr != nil {
+			m.logger.Debugf("Resource request CRUD Error: %v\n", requestErr)
+			http.Error(w, fmt.Sprintf("Error: %v", requestErr), http.StatusBadRequest)
+		}
+	}()
+
+	if r.Method != "POST" {
+		requestErr = fmt.Errorf("verb not supported: %v", r.Method)
+		return
+	}
+
+	id := mux.Vars(r)["id"]
+	if id == "" {
+		http.Error(w, "Var `id` must be set", http.StatusBadRequest)
+		return
+	}
+
+	newMgr, ok := m.manager.(bundle.NewManagement)
+	if !ok {
+		serverErr = errors.New("server does not support resource CRUD operations")
+		return
+	}
+
+	ctx, done := context.WithDeadline(r.Context(), time.Now().Add(m.apiTimeout))
+	defer done()
+
+	var storeFn func(*yaml.Node)
+
+	docType := docs.Type(mux.Vars(r)["type"])
+	switch docType {
+	case docs.TypeCache:
+		storeFn = func(n *yaml.Node) {
+			cacheConf := cache.NewConfig()
+			if requestErr = n.Decode(&cacheConf); requestErr != nil {
+				return
+			}
+			serverErr = newMgr.StoreCache(ctx, id, cacheConf)
+		}
+	case docs.TypeInput:
+		storeFn = func(n *yaml.Node) {
+			inputConf := input.NewConfig()
+			if requestErr = n.Decode(&inputConf); requestErr != nil {
+				return
+			}
+			serverErr = newMgr.StoreInput(ctx, id, inputConf)
+		}
+	case docs.TypeOutput:
+		storeFn = func(n *yaml.Node) {
+			outputConf := output.NewConfig()
+			if requestErr = n.Decode(&outputConf); requestErr != nil {
+				return
+			}
+			serverErr = newMgr.StoreOutput(ctx, id, outputConf)
+		}
+	case docs.TypeProcessor:
+		storeFn = func(n *yaml.Node) {
+			procConf := processor.NewConfig()
+			if requestErr = n.Decode(&procConf); requestErr != nil {
+				return
+			}
+			serverErr = newMgr.StoreProcessor(ctx, id, procConf)
+		}
+	case docs.TypeRateLimit:
+		storeFn = func(n *yaml.Node) {
+			rlConf := ratelimit.NewConfig()
+			if requestErr = n.Decode(&rlConf); requestErr != nil {
+				return
+			}
+			serverErr = newMgr.StoreRateLimit(ctx, id, rlConf)
+		}
+	default:
+		http.Error(w, "Var `type` must be set to one of `cache`, `input`, `output`, `processor` or `rate_limit`", http.StatusBadRequest)
+		return
+	}
+
+	var confNode *yaml.Node
+	var lints []string
+	{
+		var confBytes []byte
+		if confBytes, requestErr = ioutil.ReadAll(r.Body); requestErr != nil {
+			return
+		}
+		confBytes = text.ReplaceEnvVariables(confBytes)
+
+		var node yaml.Node
+		if requestErr = yaml.Unmarshal(confBytes, &node); requestErr != nil {
+			return
+		}
+		confNode = &node
+
+		if r.URL.Query().Get("chilled") != "true" {
+			for _, l := range docs.LintNode(docs.NewLintContext(), docType, &node) {
+				lints = append(lints, fmt.Sprintf("line %v: %v", l.Line, l.What))
+				m.logger.Infof("Resource '%v' config: %v\n", id, l)
+			}
+		}
+	}
+	if len(lints) > 0 {
+		errBytes, _ := json.Marshal(struct {
+			LintErrs []string `json:"lint_errors"`
+		}{
+			LintErrs: lints,
+		})
+		w.WriteHeader(http.StatusBadRequest)
+		w.Write(errBytes)
+		return
+	}
+
+	storeFn(confNode)
 }
 
 // HandleStreamStats is an http.HandleFunc for obtaining metrics for a stream.