+++ CHANGELOG.md
@@ -28,6 +28,7 @@ All notable changes to this project will be documented in this file.
 - New Bloblang method `parse_yaml`.
 - CLI flag `env-file` added to Benthos for parsing dotenv files.
 - New `mssql` SQL driver for the `sql` processor and output.
+- New POST endpoint `/resources/{type}/{id}` added to Benthos streams mode for dynamically mutating resource configs.
 
 ### Changed